+++ lib/core/StateMachine.cpp
