+++ lib/core/actions/SequentialAction.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Action.h"
-#include "CommandGenerator.h"
 #include "Commands.h"
 #include <memory>
 #include <vector>
