+++ lib/io-teensy/EncoderSensor.h
