+++ lib/io-teensy/Gyro.h
@@ -1,0 +1,21 @@
+#pragma once
+#include "MPU6050_6Axis_MotionApps20.h"
+#include "Wire.h"
+
+struct Gyro {
+  MPU6050 mpu;
+  bool dmpReady = false;
+  uint8_t devStatus = 0;
+  uint16_t packetSize = 0;
+  uint8_t fifoBuffer[64];
+
+  float ypr[3] = {0, 0, 0};
+
+  Quaternion q;
+  VectorFloat gravity;
+
+  void initalizeGyro();
+
+  // returns true when new ypr was computed
+  bool update();
+};
