+++ platformio.ini
@@ -18,10 +18,6 @@ monitor_speed = 9600
 build_unflags = -std=gnu++11
 build_flags =
     -std=gnu++20
-    -Isrc
-    -Isrc/include
-    -Isrc/actions
-    -Isrc/MouseIO
 lib_deps =
   jrowberg/I2Cdevlib-Core
   jrowberg/I2Cdevlib-MPU6050
