--- src/include/FastPathSolver.h
@@ -1,14 +1,0 @@
-#pragma once
-#include "CellSelection.h"
-#include "CommandAction.h"
-#include "CommandGenerator.h"
-#include "Solver.h"
-#include <vector>
-
-class FastPathSolver final : public Solver {
-public:
-  Action *run(MouseState &state, const Goals *goal) override;
-  bool end(MouseState &state, const Goals *goal) override;
-  void onFinished(MouseState &state, const Goals *goal) override;
-  void init(MouseState &state, const Goals *goal) override;
-};
